feat: Unified Helper classes into a single one
